+++ streampipes-extensions/streampipes-sinks-internal-jvm/src/main/java/org/apache/streampipes/sinks/internal/jvm/InternalSinksExtensionModuleExports.java
@@ -24,7 +24,7 @@ import org.apache.streampipes.extensions.api.migration.IModelMigrator;
 import org.apache.streampipes.extensions.api.pe.IStreamPipesPipelineElement;
 import org.apache.streampipes.sinks.internal.jvm.datalake.DataLakeSink;
 import org.apache.streampipes.sinks.internal.jvm.datalake.migrations.DataLakeSinkMigrationV1;
-import org.apache.streampipes.sinks.internal.jvm.notification.StreamPipesNotificationSink;
+import org.apache.streampipes.sinks.internal.jvm.notification.InternalStreamPipesNotificationSink;
 
 import java.util.Collections;
 import java.util.List;
@@ -39,7 +39,7 @@ public class InternalSinksExtensionModuleExports implements IExtensionModuleExpo
   public List<IStreamPipesPipelineElement<?>> pipelineElements() {
     return List.of(
         new DataLakeSink(),
-        new StreamPipesNotificationSink()
+        new InternalStreamPipesNotificationSink()
     );
   }
