+++ streampipes-extensions/streampipes-sinks-notifications-jvm/pom.xml
@@ -41,11 +41,6 @@
         </dependency>
         <dependency>
             <groupId>org.apache.streampipes</groupId>
-            <artifactId>streampipes-sinks-internal-jvm</artifactId>
-            <version>0.95.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.streampipes</groupId>
             <artifactId>streampipes-wrapper-standalone</artifactId>
             <version>0.95.0-SNAPSHOT</version>
         </dependency>
